Dedupe boundwith references when both forms appear
When the boundwith references are populated both as linkages in the holding records and as empty item records, only the empty item records should be represented in the OPAC*. If the two references lead to different status information, the more recent status is shown**.

DISCOVERYACCESS-3531

* Suppression depends on the number of boundwith holding references matching the number of empty item records.

** Currently, this is only for single volume situations. When multiple boundwith volumes are in play, the empty item status is used. In the future, we could attempt to match the two sets in order to make an item-by-item status comparison.

+++ src/test/java/edu/cornell/library/integration/voyager/PopulateTestVoyagerDB.java
@@ -31,12 +31,12 @@ public class PopulateTestVoyagerDB {
 
   public static boolean replaceDBContents = false; // if false, will add specified bibs to existing tables
   public static String testDbConnectionString = "jdbc:sqlite:src/test/resources/voyagerTest.db";
-  private static List<Integer> testBibs = Arrays.asList(369282);
+  private static List<Integer> testBibs = Arrays.asList(833840,836782);
 
   // Bibs in "jdbc:sqlite:src/test/resources/voyagerTest.db"
   // 330581,3212531,2248567,576519,3827392,1016847,969430,1799377,2095674,1575369,9520154,927983,
   // 342724,4442869,784908,6047653,9628566,3956404,9647384,306998,329763,2026746,4546769,10023626
-  // 867,9295667,1282748,4888514,369282
+  // 867,9295667,1282748,4888514,369282,833840,836782
 
   private static boolean dumpTestDBToStdout = false;
   private static Set<Integer> testMfhds = new HashSet<>();
@@ -528,10 +528,10 @@ public class PopulateTestVoyagerDB {
   }
 
   private static void line_item(Connection testDb, Connection voyager, Statement testDbstmt) throws SQLException {
-//    if (replaceDBContents) {
+    if (replaceDBContents) {
       testDbstmt.executeUpdate("drop table if exists line_item");
       testDbstmt.executeUpdate("create table line_item ( line_item_id int, bib_id int )");
- //   }
+    }
     System.out.println("Loading line_item data for "+testBibs.size()+" bibs.");
     try (PreparedStatement readStmt = voyager.prepareStatement(
         "SELECT line_item_id FROM line_item WHERE bib_id = ?");
@@ -553,10 +553,10 @@ public class PopulateTestVoyagerDB {
   }
 
   private static void line_item_copy_status(Connection testDb, Connection voyager, Statement testDbstmt) throws SQLException {
-//    if (replaceDBContents) {
+    if (replaceDBContents) {
       testDbstmt.executeUpdate("drop table if exists line_item_copy_status");
       testDbstmt.executeUpdate("create table line_item_copy_status ( line_item_id int, mfhd_id int )");
-//    }
+    }
     System.out.println("Loading line_item_copy_status data for "+testMfhds.size()+" mfhds.");
     try (PreparedStatement readStmt = voyager.prepareStatement(
         "SELECT line_item_id FROM line_item_copy_status WHERE mfhd_id = ?");
@@ -578,10 +578,10 @@ public class PopulateTestVoyagerDB {
   }
 
   private static void subscriptionAndIssues(Connection testDb, Connection voyager, Statement testDbstmt) throws SQLException {
-//    if (replaceDBContents) {
+    if (replaceDBContents) {
       testDbstmt.executeUpdate("drop table if exists subscription");
       testDbstmt.executeUpdate("create table subscription ( line_item_id int, subscription_id int )");
-//    }
+    }
     System.out.println("Loading subscription data for "+testLineItems.size()+" line_item_ids.");
     try (PreparedStatement readStmt = voyager.prepareStatement(
         "SELECT subscription_id FROM subscription WHERE line_item_id = ?");
@@ -604,10 +604,10 @@ public class PopulateTestVoyagerDB {
   
   private static void component(Connection testDb, Connection voyager, Statement testDbstmt) throws SQLException {
 
-//    if (replaceDBContents) {
+    if (replaceDBContents) {
       testDbstmt.executeUpdate("drop table if exists component");
       testDbstmt.executeUpdate("create table component ( component_id int, subscription_id int )");
-//    }
+    }
     System.out.println("Loading component data for "+subscriptions.size()+" subscriptions.");
     try (PreparedStatement readStmt = voyager.prepareStatement(
         "SELECT component_id FROM component WHERE subscription_id = ?");
@@ -631,10 +631,10 @@ public class PopulateTestVoyagerDB {
 
   private static void issues_received(Connection testDb, Connection voyager, Statement testDbstmt) throws SQLException {
 
-//    if (replaceDBContents) {
+    if (replaceDBContents) {
       testDbstmt.executeUpdate("drop table if exists issues_received");
       testDbstmt.executeUpdate("create table issues_received ( component_id int, issue_id int, opac_suppressed int )");
-//    }
+    }
     System.out.println("Loading issues_received data for "+components.size()+" components.");
     try (PreparedStatement readStmt = voyager.prepareStatement(
         "SELECT issue_id, opac_suppressed FROM issues_received WHERE component_id = ?");
@@ -657,10 +657,10 @@ public class PopulateTestVoyagerDB {
 
   private static void serial_issues(Connection testDb, Connection voyager, Statement testDbstmt) throws SQLException {
 
-//    if (replaceDBContents) {
+    if (replaceDBContents) {
       testDbstmt.executeUpdate("drop table if exists serial_issues");
       testDbstmt.executeUpdate("create table serial_issues ( component_id int, issue_id int, enumchron string, receipt_date date )");
-//    }
+    }
     System.out.println("Loading serial_issues data for "+components.size()+" components.");
     try (PreparedStatement readStmt = voyager.prepareStatement(
         "SELECT issue_id, enumchron, receipt_date FROM serial_issues WHERE component_id = ?");
